Change Shape Square and Rectangle classes to allow for inhertitance call

+++ exercises/Samuel_Jones/ex02_oo_basics/src/Rectangle.cpp
@@ -4,10 +4,10 @@ Rectangle::Rectangle(int side1, int side2):m_side1(side1), m_side2(side2), Shape
 {
 }
 
-int Rectangle::getPerimeter(){
+float Rectangle::getPerimeter(){
     return m_side1*2 + m_side2*2;
 }
 
-int Rectangle::getArea(){
+float Rectangle::getArea(){
     return m_side1*m_side2;
 }
